Issue #15: Shutdown Command Hanging on Execution
As it turned out, the shutdown command was getting invalid results returned from the check for admin, resulting from poorly formed participant data.
The issue here isn't the command itself, but that participant data needs an overhaul. For the moment, I have remarked out the check for admin so that
we can continue testing and development.