Issue #48: Help Command Should Address Logout Command
Final kinks ironed out - there was a little problem with the argument cluster
not counting arguments correctly, and that was making the help command occasionally
segfault.

That's fixed now, and the help command finally reliably addresses the logout command.

Lots of foundation work here for the other commands as well, so we should see those
start getting knocked off the project in rapid order.