Issue #367: Connection and Participant Data Unification - Project Branch Issue #390: Scenario Passthrough: Egg Attempts to Login While Already Logged In
get.participant.data.by.name was looking to
answer.active.socket.packet.bsize for the size of the retrieved
data, which was a correct assumption given that recv.from.socket
was putting that data there at the time. Unfortunately,
answer.active.socket.packet.bsize is kind of a high traffic field,
and was getting clobbered on the regular, so now we use
answer.active.socket.packet.rsize instead.