Issue #367: Connection and Participant Data Unification - Project Branch Issue #390: Scenario Passthrough: Egg Attempts to Login While Already Logged In
Complete revamp of clean.socket. The code that was there before
was more or less a placeholder, and had been almost completely
obsoleted by changes to the architecture. Using
answer.active.socket.packet instead of it's own structure, and now
makes proper differentiation between the types of credentials with
which it's working.