Issue #367: Connection and Participant Data Unification - Project Branch Issue #391: Scenario Passthrough: Egg Logs In For the First Time and Sets Passphrase
I realized that the reason I wasn't seeing any further activity
from the login after giving it a name was that in fact, the socket
was now operating out of an egg credential. Boneheaded. Egg
polling in active.socket.poll was still pretty primitive. This
commit shores it up and makes it functional.