Issue #367: Connection and Participant Data Unification - Project Branch Issue #392: Scenario Passthrough: Old Egg Logs In and Uses Correct Passphrase
As it turns out, answer.active.socket.packet.fpointer needs to be
correctly populated before it can be used by status handlers to
open credential files. That happens now in answer.active.socket.

This will probably change later on - I'm looking at all the
diversions in the execution path to make determinations about
which type of socket we're looking at, and I'm thinking that
distinction should only be made once, preferably as early on as
possible.