Issue #367: Connection and Participant Data Unification - Project Branch Issue #392: Scenario Passthrough: Old Egg Logs In and Uses Correct Passphrase
Reworked what was essentially placeholder code for
address.passphrase.status. The old code relied upon a unique
structure, which was abandoned for the new architecture, and was
not aware of a number of new changes to the way some of it's own
dependencies worked.

This has all been fixed, and now old eggs may log in with correct
passphrases, and be given access to the environment as well as the
command shell.